first hjcommit
