+++ .env
@@ -1,7 +1,7 @@
 DATABASE_HOST = localhost
 DATABASE_USER = root
-DATABASE_PASSWORD = 0000
-DATABASE = test
+DATABASE_PASSWORD = 1234
+DATABASE = sys
